fix: align model loading and bnb config with fp16 for T4 compatibility

+++ paradox_v2/training/train_qlora.py
@@ -60,11 +60,11 @@ def train(base_model=None, dataset_path=None, output_dir=None,
     print("[2/5] Loading model with 4-bit NF4...")
     bnb_config = BitsAndBytesConfig(
         load_in_4bit=True, bnb_4bit_quant_type="nf4",
-        bnb_4bit_compute_dtype=torch.bfloat16, bnb_4bit_use_double_quant=True,
+        bnb_4bit_compute_dtype=torch.float16, bnb_4bit_use_double_quant=True,
     )
     model = AutoModelForCausalLM.from_pretrained(
         base_model, quantization_config=bnb_config, device_map="auto",
-        trust_remote_code=True, torch_dtype=torch.bfloat16,
+        trust_remote_code=True, torch_dtype=torch.float16,
     )
     model = prepare_model_for_kbit_training(model)
